Providing MSVC alternatives to assembly code
The MSVC compiler does not support inline assembly. The alternative is
to use intrinsic functions, which are available through a number of
header files. This affects two places in our code: the initial query to
the CPU to check whether SSE4.2 is supported, and the crc32c function
itself.

In the case of the cpuid instruction, the function wrapping it quite
simple, so we provide two different implementations (i.e., one for MSVC,
one for the other compilers).

In the case of the crc32c instruction, and because it involves a bit
more of logic, we provide a single implementation that is heavily
typedef'd and macro'd so it works across compilers and architectures.
In particular, the different intrinsic functions provided by MSVC have
slightly different signatures from one to another (the type of the crc
value is different depending on the instruction), so we need to perform
a few pointer and value casts to make them properly work.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>