Making code C89-aware
This requirement only comes from the fact that we are trying to support
MSVC, which until 2015 apparently didn't implement C99 fully yet and
therefore defaulted to the C89 standard.

These adjustments include using C89-style variable declaration, and the
lack of the stdint.h header file. The latter was used to pull the
definition of uint32_t and uint64_t, so in case we don't find it we
manually define these types.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>