New software implementation of crc32c
We don't automatically fallback to this implementation when no hardware
support is found. Instead, the user has to explicitly request this
package to be loaded in "sw mode" via an environment variable. The
rationale for this is that we don't want to give a false sense of
security to people: if the package always successfully imported in
all platforms users would think their CPU supports the SSE4.2
extensions, which might not be true.

The software implementation is based on that written by Mark Adler
(search the web, it's an easy find), which is released under the BSD
3-clause license. Some small modifications were needed to make it
happily compile on Python under different platforms. On the one hand,
variables cannot be declared in 'for' statements (because python doesn't
necessarily compile the code with C99). On the other hand, the code was
unnecessarily using the int32_t type (which is not defined in MSVC)
inside a sizeof() call. We can safely use the uint32_t type instead, as
it has the same size and is already defined (by us) for all MSVC
versions.

We are running the unit tests for this software-based method now too to
ensure it's working like the hardware-based method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>