Load result from unsigned long
The crc value is handled always as an uint32_t, but only until we give
it to Python to convert it into an object, in which case it was
interpreted as a long (via PyInt_fromLong). This produced the effect
that to compare values at the python level one needed to AND the result
with 0xffffffff; otherwise values with their highest bit set would be
interpreted as negative values and would fail to compare equally to
their bit-identical positive counterparts.

To avoid this problem we now always use PyLong_FromUnsignedLong, which
takes an unsigned long as an argument, and therefore properly interprets
the value as unsigned.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>