crc32c 1.4
This version supports Windows builds via the Visual Studio compiler. To
ensure this is indeed working we have added both a set of unit tests,
and AppVeyor/Travis build scripts.

This version also drops the old consumption method, which
read/crc'd/wrote in a tight loop. This was always an experimental
feature so there is no harm on taking it out.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>