Producing and upload manylinux binary wheels
To achieve this we are using the `manylinux` container, which is based
on Centos 5 and therefore gives us the best compatibility across old and
new Linux distributions. Our package has no external dependencies, so
it's as easy as producing the weel for each flavour of Linux and then
tagging it as `manylinux`1 with auditwheel (to make sure we actually can).

Instead of individually doing this on each Travis sub-job, we create a
new one that does *only* this. After producing the wheels it also
uploads them to PyPI.

The whole process of creating and uploading wheels in the new sub-job is
only run if the current commit is a tag; otherwise we don't do any of it
and quickly return.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>