Support for compiling with icc
icc, like msvc, doesn't have the built-in functions provided by
gcc/clang, but it does have the intrinsic functions defined in the
intrinsic headers.

Food for thought: maybe it would be simpler to use the intrinsics
instead of the built-ins? So far this is working, so let's leave it as
it is for the moment.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>