Allowing CRC32C_SW_MODE=auto/force values
'force' has the same meaning that the current '1' value; on the other
hand 'auto' means that the software implementation should be used as a
fallback when no hardware support is found. This is useful for cases
when crc32c support needs to be provided regardless of the underlying
implementation (and therefore speed).

Since we are already deprecating '1' I'm setting the example and
am updating the travis/appveyor scripts to use 'force' instead.

This fixes #5.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>