Fix using build machine compiler flags rather than target
This fix uses the 'setup.cfg' file used by setuptools to define parameters
for cross-compiling.

The reason for using this config file is that command-line parameters can't
be passed to setup.py, as the command-line is ideally opaque to the
installer. 'setup.cfg' was added to setuptools as a useful middle-ground
between modifying 'setup.py' and the command line. Now, build systems can
modify a config file instead of the source directly.