New default behavior when hw support is missing
The previous behavior raised an ImportError when no hardware support was
found by default, unless the user opted into using the software support.
With time it became clear that this was not the optimal solution, as
most people simply want to use the algorithm, regardless of software
support.

With this change the default behavior is to fallback into using the
software implementation of the crc32c algorithm instead of failing.
Users can still opt out of this by specified CRC32C_SW_MODE=none, which
will give them the old behavior.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>