crc32c v2.0
This new version changes the import logic to automatically fallback to
using the software implementation when no hardware support can be found
(and hence the package doesn't fail to import).

Since this is a change in behavior, a new major version is being
produced. Users shouldn't have many problems though.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>