Declare all Extension sources unconditionally
The dynamic nature in which sources were added to the Extension object
at build time meant that the extra sources (crc32_adler.c and
checksse42.c) were not included in source distributions produced via the
"sdist" setup.py command (because the "build" command does not run as
part of "sdist"). This in turn meant that trying to install such source
distributions in Intel platforms failed because the building process
failed to find the missing files in the source distributions.

This problem was introduced in 21a464c but went unnoticed until now
because we have been producing source distributions locally using the
exact same environment for years, before that particular change was
introduced (and therefore our local copy of the MANIFEST already
contained the two files as part of the distribution). Now that we are
moving towards producing all distribution files in GitHub via its
Actions functionality we stumbled upon this particular issue.

Fixing this problem involves adding back the two source files to be part
of the Extension unconditionally, and instead add the proper macro
checks to include the file's contents conditionally depending on the
target platform.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>