Reinstate Travis configuration file
Travis offers arm64 images, which is currently not offered by GitHub
actions, and therefore there is still value in using Travis to ensure
that our package compiles successfully in arm64 platforms.

This is a much simpler version of what we previously had. In 875532e we
a single python script was produced that runs all the necessary tests
internally. We are running that script as the only test now in Travis,
and also expanding the job matrix to include more python versions. We
also run the installation step more verbosely.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>