Add first version of ARM support checks
Checking for CRC32C instruction support in ARM is not something that can
be done from userspace, like in the case of Intel CPUs. Instead, the
kernel must provide this information to user-space programs. This is
done in the case of Linux through the "auxiliary value", which programs
can query through the Linux-specific getauxval call.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>