Deprecate crc32 in favour of crc32c
The existing implementation of the crc32c algorithm already does what we
need it to do, but the function exposing this functionality is wrongly
named (in order to be as close as possible to binascii's crc32 function,
even at the name level). This has no advantage though, and only helps
confusing people.

This commit adds a new function in the C module that first issues a
warning about *it* being deprecated, and then immediately calls our
original function. The new function is then named "crc32", while the old
one is renamed to "crc32c", completing the renaming cycle.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>