Add missing LICENCE and AUTHOR files
This project borrows code from several locations, but we were not
properly including their licenses here, even though the licenses
themselves state that they should be reproduced.

The missing licenses were the 3-clause BSD licenses from MIT and
google's crc32c. The original slice-by-8 project didn't include a
LICENSE file, only a link to the 2-clause BSD license, so we'll do the
same.

Additionally, google's crc32c contains an AUTHOR file for copyright
reasons, so we are added that as well (as it was at the moment of
including some of their code in our package).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>