Further clarify licensing/copyrighting of code
Before we had a very short explanation on how the different parts of the
code were licensed/copyrighted, but it was brought to our attention that
further clarifications were required. Now that we have added the missing
LICENSE/AUTHOR files from the projects we have borrowed code from, we
can also tell a better and more precise story, and add links to the
correct resources.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>