Expose system endianess in new module constant
This information useful both internally and externally. Internally we
will need to act differently when calculating the crc32c value (via the
software implementation), and externally users will need to know the
endianess of their system (which could probably be obtained somehow
else, but we're good citizens) if they want to get an interoperable byte
sequence for their checksum values.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>