Add big endian crc32c software implementation
Since our implementation does multiple 4-byte values at a time, we need
to take into account the layout in memory of those values. So far we
have worked with little-endian machines, but results will be different
for big-endian machines, where the algorithm needs to be inverted when
going through multibyte values; otherwise the algorithm it the same.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>