Move C sources to src/ subdirectory
Initially we had only one or two sources, but with time the number of
files we compile has continued growing, so it makes sense to put them
inside their own directory.

I took the change to simplify the way we declare these files on the
Extension declaration, using glob instead of an explicit list.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>