Stop forcing target flags globally while building
Remove the code appending options needed for hardware CRC32c
implementations globally.  This is incorrect, as it permits the compiler
to emit target-specific code (i.e. SSE4.2 code or ARMv8 code) on systems
where the processor does not actually support them.

This logic has been replaced by inline use of target-specific attributes
and pragmas in the individual files.  They should work reliably on x86
targets and on the majority of ARM64 targets.

Fixes #30