Never purposefully fail to import crc32c
This behaviour was still lingering from the old 1.x series, so let's
remove it. Instead of failing at import time, we can still choose to
fail at runtime, which is a better design, since it makes importing the
module much less cumbersome. And while it's a backwards-incompatible
change, nobody will miss it (and if anyone does, they can easily
replicate it on their end).

The biggest repercussion this change had was in our own tests, which
relied on the import working or not to skip tests conditionally. This
logic has now been updated, and the supporting code has been redesigned
as fixtures and markers, more in line with pytest's (and therefore, many
users and readers) expectations.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>