Don't include crc32c/ext as package data
On edacda4 we added type hint stub files that needed to be distributed
with the source and binary distributions of our package. However we
relied on the include_package_data, which recursively includes all files
it finds, and therefore we ended up with the ext/* sources included in
our binary wheels.

This commit removes the include_package_data flag usage, and replaces it
with a more details package_data specification. This means we can also
remove the corresponding lines from the MANIFEST.in file. Additionally,
the header files don't need to be specified in MANIFEST.in since they
are already stated as a dependency of the C Extension.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>