Introduce isort linting
This commit adds isort to the list of development requirements, and
introduces the required changes so that isort checks successfully our
repository. We were already doing a pretty good job, with only some
minor corrections that were required. The lint CI job has also been
updated to run this check.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>