Store crc_fn pointer in module state
We previously stored this in a global, static variable, which prevents
the module from supporting sub-interpreters. If we want to be nice
citizens we need to use the module's state memory to store this instead.

This commit adds a new CRC32CState structure (and associated utility
function) currently only holding the crc_fn pointer. The state is used
to store and later retrieve the pointer value.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>