Introduce tox
Although I was initially reluctant to use tox, two recent things have
convinced me to give it a try: firstly, I've seen now a few projects
using them to great effect; and secondly, the ability to use the tox-uv
plug-in means that it's very easy to run tests across multiple versions
of python without needing them pre-installed in the host machine.

This first version of tox declares the default environment command to be
running the unit tests, and the default list of environments to run to
be all python versions crc32c supports. It additionally declares an
environment to run the benchmarks on, and another to run the linting
checks, both of which rely on declaring the dependencies in the
pyproject.toml file itself (removing the need to maintain a separate
requirements-dev.txt file).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>