Let setuptools auto-discover src/ layout
Now that we use pyproject.toml we can use setuptools' automatic layout
discovery for our project (the so-called "src-layout"). This works fine
even without a pyproject-based configuration for pure python projects,
but those declaring extensions *need* to be configured through pyproject
according to setuptools' documentation.

Note that the "build" tool was already complaining about the
"crc32c.ext" package not being explicitly included (or excluded). We
don't really expect it, but we don't really care if it's there either.
Either way, it's easier and better to let setuptools do the src/
automatic discovery and identify packages, etc.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>